requirements.txt
speedtest-cli
Pillow